requirements.txt
streamlit
fpdf
python-dotenv
pytest
streamlit_js_eval
st-theme
